refactor: convention에 맞게 코드정리 및 import 정리
