+++ webapp/cart/admin/cartAdmin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from cart.models import Cart
+from cart.models import cart
 from survey.models import Survey
