+++ webapp/cart/models/__init__.py
@@ -1,0 +1,1 @@
+from .cart import Cart
