+++ webapp/template/views/templateViewSet.py
@@ -7,5 +7,3 @@ from template.serializers import TemplateSerializer
 class TemplateViewSet(viewsets.ModelViewSet):
     queryset = Template.objects.all()
     serializer_class = TemplateSerializer
-
-
