+++ webapp/user/views/userAPIView.py
@@ -1,8 +1,9 @@
 from rest_framework.generics import RetrieveUpdateAPIView
-from rest_framework.views import APIView
 from rest_framework.response import Response
+
 from user.models import User
-from user.serializers.UserSerializer import UpdateUserSerializer,UserViewSerializer
+from user.serializers.UserSerializer import UpdateUserSerializer
+
 
 class UpdateUserAPIView(RetrieveUpdateAPIView):
     queryset = User.objects.all()
@@ -17,4 +18,4 @@ class UpdateUserAPIView(RetrieveUpdateAPIView):
         serializer.is_valid(raise_exception=True)
         self.perform_update(serializer)
         return Response(serializer.data)
-        
+
